# routes/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample YoYo Gifts Server.  You can find out more about     Yoyo Gifts at [https://yoyo-gift-qa.web.app](https://yoyo-gift-qa.web.app) or on [irc.freenode.net, #swagger](https://yoyo-gift-qa.web.app/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "1.0.0",
    "title": "YO YO Gifts",
    "termsOfService": "https://yoyo-gift-qa.web.app/terms/",
    "contact": {
      "email": "apiteam@yoyo-gift.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "yoyo-gift-qa.web.app",
...
  "schemes": "https",
  "paths": {
...
    "/products/change/product/{_id}": {
      "put": {
        "tags": ["product"],
        "summary": "Update an existing product",
        "description": "",
        "operationId": "updateproduct",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "description": "product that need to be updated",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "product that need to be updated",
            "required": true,
            "type": "string"
          },
          {
            "name": "isAdmin",
            "in": "header",
            "description": "product that need to be updated",
            "required": true,
            "type": "boolean"
          },
          {
            "name": "body",
            "in": "body",

            "required": true,
            "schema": {
              "$ref": "#/definitions/product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/productresponse"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "product not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "productstore_auth": ["write:products", "read:products"]
          }
        ]
      }
    },
...
    "/cart/addInCart": {
      "post": {
        "tags": ["Cart"],
        "summary": "Add a new product in cart",
        "description": "registered user call this",
        "operationId": "addInCart",
        "consumes": "application/json",
        "produces": "application/json",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization token required",
            "required": true,
            "default": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InRlc3RAdGVzdC5jb20iLCJpc0FkbWluIjpmYWxzZSwiaWF0IjoxNTg3Njk5NTk0fQ.v-28iA9CRSFPbgln7_OiUoLzeCccQWs5fYmcg3FuvxU",
            "type": "string"
          },
          {
            "name": "isAdmin",
            "in": "header",
            "description": "A flag to check type of user",
            "default": false,
            "required": true,
            "type": "boolean"
          },
          {
            "in": "body",
            "name": "body",
            "description": "cart body that needs to be added in app",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/productresponse"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          },
          "200": {
            "description": "Category Saved",
            "schema": {
              "$ref": "#/definitions/productresponse"
            }
          }
        },
        "security": [
          {
            "petstore_auth": ["write:pets", "read:pets"]
          }
        ]
      }
    },
    "/cart/removeFromCart": {
      "post": {
        "tags": ["Cart"],
        "summary": "Remove a product in cart",
        "description": "registered user call this",
        "operationId": "removeFromcart",
        "consumes": "application/json",
        "produces": "application/json",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization token required",
            "required": true,
            "default": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InRlc3RAdGVzdC5jb20iLCJpc0FkbWluIjpmYWxzZSwiaWF0IjoxNTg3Njk5NTk0fQ.v-28iA9CRSFPbgln7_OiUoLzeCccQWs5fYmcg3FuvxU",
            "type": "string"
          },
          {
            "name": "isAdmin",
            "in": "header",
            "description": "A flag to check type of user",
            "default": false,
            "required": true,
            "type": "boolean"
          },
          {
            "in": "body",
            "name": "body",
            "description": "cart body that needs to be added in app",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string",
                  "example": "5e96e76abad34e45ac8589c5"
                },
                "productId": {
                  "type": "string",
                  "example": "5e8b1f572d81de21adac301d"
                }
              }
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          },
          "200": {
            "description": "Category Saved",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/productresponse"
              }
            }
          }
        },
        "security": [
          {
            "petstore_auth": ["write:pets", "read:pets"]
          }
        ]
      }
    },
    "/cart/getCart/{userId}": {
      "get": {
        "tags": ["Cart"],
        "summary": "Finds cart by userid",
        "description": "Finds cart by userid",
        "operationId": "Finds cart by userid",
        "produces": "application/json",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization token required",
            "required": true,
            "default": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InRlc3RAdGVzdC5jb20iLCJpc0FkbWluIjpmYWxzZSwiaWF0IjoxNTg3Njk5NTk0fQ.v-28iA9CRSFPbgln7_OiUoLzeCccQWs5fYmcg3FuvxU",
            "type": "string"
          },
          {
            "name": "isAdmin",
            "in": "header",
            "description": "A flag to check type of user",
            "default": false,
            "required": true,
            "type": "boolean"
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of User",
            "required": true,
            "default": "5e9e173440d262537406e460",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/productresponse"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        },
        "security": [
          {
            "petstore_auth": ["write:pets", "read:pets"]
          }
        ]
      }
    },
    "/cart/increment/{cartId}/{userId}": {
      "get": {
        "tags": ["Cart"],
        "summary": "Increment Product Count In Cart",
        "description": "Increment Product Count In Cart",
        "operationId": "Increment Product Count In Cart",
        "produces": "application/json",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization token required",
            "required": true,
            "default": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InRlc3RAdGVzdC5jb20iLCJpc0FkbWluIjpmYWxzZSwiaWF0IjoxNTg3Njk5NTk0fQ.v-28iA9CRSFPbgln7_OiUoLzeCccQWs5fYmcg3FuvxU",
            "type": "string"
          },
          {
            "name": "isAdmin",
            "in": "header",
            "description": "A flag to check type of user",
            "default": false,
            "required": true,
            "type": "boolean"
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of User",
            "required": true,
            "default": "5e9e173440d262537406e460",
            "type": "string"
          },
          {
            "name": "cartId",
            "in": "path",
            "description": "ID cart",
            "required": true,
            "default": "5ea03d865cbc190f48e0ae2f",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/productresponse"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        },
        "security": [
          {
            "petstore_auth": ["write:pets", "read:pets"]
          }
        ]
      }
    },
    "/cart/decrement/{cartId}/{userId}": {
      "get": {
        "tags": ["Cart"],
        "summary": "Decrement Product Count In Cart",
        "description": "Decrement Product Count In Cart",
        "operationId": "Decrement Product Count In Cart",
        "produces": "application/json",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization token required",
            "required": true,
            "type": "string",
            "default": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InRlc3RAdGVzdC5jb20iLCJpc0FkbWluIjpmYWxzZSwiaWF0IjoxNTg3Njk5NTk0fQ.v-28iA9CRSFPbgln7_OiUoLzeCccQWs5fYmcg3FuvxU"
          },
          {
            "name": "isAdmin",
            "in": "header",
            "description": "A flag to check type of user",
            "default": false,
            "required": true,
            "type": "boolean"
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of User",
            "required": true,
            "default": "5e9e173440d262537406e460",
            "type": "string"
          },
          {
            "name": "cartId",
            "in": "path",
            "description": "ID cart",
            "required": true,
            "default": "5ea03d865cbc190f48e0ae2f",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/productresponse"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        },
        "security": [
          {
            "petstore_auth": ["write:pets", "read:pets"]
          }
        ]
      }
    },
    "/cart/addManyToCart": {
      "post": {
        "tags": ["Cart"],
        "summary": "Add cart items in db when user logged in",
        "description": "Add items in cart when user logged in",
        "operationId": "addManyInCart",
        "consumes": "application/json",
        "produces": "application/json",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization token required",
            "required": true,
            "type": "string"
          },
          {
            "name": "isAdmin",
            "in": "header",
            "description": "A flag to check type of user",
            "default": false,
            "required": true,
            "type": "boolean"
          },
          {
            "in": "body",
            "name": "body",
            "description": "cart body that needs to be added in app",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string",
                  "required": "true"
                },
                "cart": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/productresponse"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          },
          "200": {
            "description": "Category Saved",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/productresponse"
              }
            }
          }
        },
        "security": [
          {
            "petstore_auth": ["write:pets", "read:pets"]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "productresponse": {
      "type": "object",
      "properties": {
        "reviews": {
          "type": "string",
          "example": []
        },
        "_id": {
          "type": "string",
          "example": "22323der23123ssa231213"
        },

        "product_name": {
          "type": "string",
          "example": "watch"
        },
        "category": {
          "type": "string",
          "example": "watch"
        },

        "description": {
          "type": "string",
          "example": "great watch"
        },
        "price": {
          "type": "integer",
          "format": "int32"
        },

        "discount": {
          "type": "integer",
          "format": "int32"
        },
        "created_at": {
          "type": "integer",
          "format": "date-time"
        },
        "updated_at": {
          "type": "integer",
          "format": "date-time"
        }
      }
    },
...
    "product": {
      "type": "object",

      "properties": {
        "product_name": {
          "type": "string",
          "example": "watch"
        },
        "category": {
          "type": "string",
          "example": "22323der23123ssa231213"
        },

        "description": {
          "type": "string",
          "example": "great watch"
        },
        "price": {
          "type": "integer",
          "format": "int32"
        },

        "discount": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
  },
...
}